Use different path for each board id.
