+++ src/components/FooterBar.vue
@@ -1,0 +1,8 @@
+<template>
+    <footer class="flex flex-col max-w-screen-xl mx-auto items-center gap-4 mt-8 py-16 px-8 text-sm">
+        <div>Created with ❤️ by <a href="mailto:tom@tomgregory.com">Tom Gregory</a>.</div>
+        <RouterLink to="/terms-and-conditions">
+            Terms and conditions
+        </RouterLink>
+    </footer>
+</template>
